+++ scripts/lib/common.sh
@@ -14,10 +14,10 @@ readonly NIXOA_DEFAULT_USERNAME="nixoa"
 readonly NIXOA_DEFAULT_TIMEZONE="America/Chicago"
 readonly NIXOA_DEFAULT_GIT_NAME="NiXOA Admin"
 readonly NIXOA_DEFAULT_GIT_EMAIL="nixoa@nixoa"
-readonly NIXOA_FLAKEHUB_SUBSTITUTER="https://cache.flakehub.com"
+readonly NIXOA_DETERMINATE_SUBSTITUTER="https://install.determinate.systems"
 readonly NIXOA_XO_SUBSTITUTER="https://xen-orchestra-ce.cachix.org"
 readonly NIXOA_LIBVHDI_SUBSTITUTER="https://libvhdi-nixpkg.cachix.org"
-readonly NIXOA_FLAKEHUB_PUBLIC_KEY="cache.flakehub.com-3:hJuILl5sVK4iKm86JzgdXW12Y2Hwd5G07qKtHTOcDCM="
+readonly NIXOA_DETERMINATE_PUBLIC_KEY="cache.flakehub.com-3:hJuILl5sVK4iKm86JzgdXW12Y2Hwd5G07qKtHTOcDCM="
 readonly NIXOA_XO_PUBLIC_KEY="xen-orchestra-ce.cachix.org-1:WAOajkFLXWTaFiwMbLidlGa5kWB7Icu29eJnYbeMG7E="
 readonly NIXOA_LIBVHDI_PUBLIC_KEY="libvhdi-nixpkg.cachix.org-1:HvYHKZcfczn2nGfCmd7F21E/MDZrlaXtN3p9mWAZT/4="
 readonly -a NIXOA_TRACKED_PATHS=(
@@ -432,8 +432,8 @@ nixoa_append_first_install_nix_options() {
   local -n command_ref="$1"
   command_ref+=(
     --option extra-experimental-features "nix-command flakes"
-    --option extra-substituters "$NIXOA_FLAKEHUB_SUBSTITUTER $NIXOA_XO_SUBSTITUTER $NIXOA_LIBVHDI_SUBSTITUTER"
-    --option extra-trusted-public-keys "$NIXOA_FLAKEHUB_PUBLIC_KEY $NIXOA_XO_PUBLIC_KEY $NIXOA_LIBVHDI_PUBLIC_KEY"
+    --option extra-substituters "$NIXOA_DETERMINATE_SUBSTITUTER $NIXOA_XO_SUBSTITUTER $NIXOA_LIBVHDI_SUBSTITUTER"
+    --option extra-trusted-public-keys "$NIXOA_DETERMINATE_PUBLIC_KEY $NIXOA_XO_PUBLIC_KEY $NIXOA_LIBVHDI_PUBLIC_KEY"
   )
 }